MVC Architecture complete
