+++ app/src/main/java/com/github/manonfrnr/blabla/Constants.java
@@ -3,4 +3,5 @@ package com.github.manonfrnr.blabla;
 public class Constants {
     public static  String KEY_POKEMON_LIST = "jsonPokemonList";
     public static  String PREFERENCES_NAME = "pokemon_application";
+    public static final String BASE_URL = "https://pokeapi.co/";
 }
