+++ app/src/main/java/com/github/manonfrnr/blabla/presentation/controller/MainController.java
@@ -1,0 +1,111 @@
+package com.github.manonfrnr.blabla.presentation.controller;
+
+import android.content.Context;
+import android.content.SharedPreferences;
+import android.widget.Toast;
+
+import com.github.manonfrnr.blabla.Constants;
+import com.github.manonfrnr.blabla.data.PokeApi;
+import com.github.manonfrnr.blabla.presentation.model.Pokemon;
+import com.github.manonfrnr.blabla.presentation.model.RestPokemonResponse;
+import com.github.manonfrnr.blabla.presentation.view.MainActivity;
+import com.google.gson.Gson;
+import com.google.gson.GsonBuilder;
+import com.google.gson.reflect.TypeToken;
+
+import java.lang.reflect.Type;
+import java.util.List;
+
+import retrofit2.Call;
+import retrofit2.Callback;
+import retrofit2.Response;
+import retrofit2.Retrofit;
+import retrofit2.converter.gson.GsonConverterFactory;
+
+public class MainController {
+
+    private SharedPreferences sharedPreferences;
+    private Gson gson;
+    private MainActivity view;
+
+    public MainController( MainActivity mainActivity, Gson gson, SharedPreferences sharedPreferences){
+        this.view = mainActivity;
+        this.gson = gson;
+        this.sharedPreferences = sharedPreferences;
+    }
+
+    public void onStart(){
+        List<Pokemon> pokemonList = getDataFromCache();
+        if(pokemonList != null){
+            view.showList(pokemonList);
+        } else{
+            makeApiCall();
+        }
+    }
+
+    private void makeApiCall() {
+
+        Retrofit retrofit = new Retrofit.Builder()
+                .baseUrl(Constants.BASE_URL)
+                .addConverterFactory(GsonConverterFactory.create(gson))
+                .build();
+
+        PokeApi pokeApi = retrofit.create(PokeApi.class);
+
+        Call<RestPokemonResponse> call = pokeApi.getPokemonResponse();
+        call.enqueue(new Callback<RestPokemonResponse>() {
+            @Override
+            public void onResponse(Call<RestPokemonResponse> call, Response<RestPokemonResponse> response) {
+                if (response.isSuccessful() && response.body() != null) {
+                    List<Pokemon> pokemonList = response.body().getResults();
+                    //Toast.makeText(getApplicationContext(), "API Success", Toast.LENGTH_SHORT).show();
+                    saveList(pokemonList);
+                    view.showList(pokemonList);
+                } else {
+                    view.showError();
+                }
+            }
+
+            @Override
+            public void onFailure(Call<RestPokemonResponse> call, Throwable t) {
+                view.showError();
+
+            }
+        });
+    }
+
+    private List<Pokemon> getDataFromCache() {
+        String jsonPokemon = sharedPreferences.getString(Constants.KEY_POKEMON_LIST, null);
+        if (jsonPokemon == null) {
+            return null;
+        } else {
+            Type listType = new TypeToken<List<Pokemon>>(){}.getType();
+            return gson.fromJson(jsonPokemon, listType);
+        }
+    }
+
+    private void saveList(List<Pokemon> pokemonList) {
+        String jsonString = gson.toJson(pokemonList);
+        sharedPreferences
+                .edit()
+                .putInt("cle_integer", 3)
+                .putString(Constants.KEY_POKEMON_LIST, jsonString)
+                .apply();
+
+        //Toast.makeText(getApplicationContext(), "List Saved", Toast.LENGTH_SHORT).show();
+
+    }
+
+    public void onItemClick(Pokemon pokemon){
+
+    }
+
+    public void onButtonAClick(){
+
+    }
+
+    public void onButtonBClick(){
+
+    }
+
+}
